# backend/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internationalized culture / e-commerce / custom-travel platform for Jingdezhen ceramic art.\nBackend in Go + Fiber; money as BIGINT minor units (CNY base, USD/EUR/GBP presentment).",
        "title": "Jingdezhen Ceramics Platform API",
        "termsOfService": "https://jingdezhen.test/terms",
        "contact": {
            "name": "Jingdezhen Ceramics Platform Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/artists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated list of artists filtered by locale + status. Access: content_editor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "artists"
                ],
                "summary": "List artists (admin, any status)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "BCP 47 locale",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by workflow status (draft|in_review|published|rejected)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.Artist"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid locale",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (needs content.write)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates an artist + its first translation. Locale defaults to en-US.\nAccess: content_editor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "artists"
                ],
                "summary": "Create an artist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Artist to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.CreateArtistData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.Artist"
                        }
                    },
                    "400": {
                        "description": "Invalid body / validation / bad locale",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (needs content.write)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/artists/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transitions an in_review artist translation to published.\nAccess: super_admin ONLY (content.publish). Body: {locale}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "artists",
                    "workflow"
                ],
                "summary": "Approve + publish an artist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "{locale: en-US}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.Artist"
                        }
                    },
                    "400": {
                        "description": "Invalid artist ID / body / bad locale",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (needs content.publish — super_admin only)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Artist not found",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Invalid workflow transition",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/artists/{slug}": {
            "get": {
                "description": "Fetches a single published artist by its locale-specific slug.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog",
                    "artists"
                ],
                "summary": "Get an artist by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist slug (locale-specific)",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"en-US\"",
                        "description": "BCP 47 locale (e.g. en-US). Overrides Accept-Language.",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.Artist"
                        }
                    },
                    "400": {
                        "description": "Missing slug",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Artist not found (or not published in this locale)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jingdezhen-ceramics-backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Optional additional details",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccess_token\u003e\" — JWT access token from POST /auth/login (or OAuth). Required for all /admin/* and customer-authed endpoints; omitted for public /catalog/* reads.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}